# core/vm: Add method Stack.popptr and apply to a bunch of instructions