… core/vm/instructions.go