@fjl 2022-05-11
This is a nice optimization idea, but very dangerous. `popptr` returns a pointer to the 'inaccessible' area of the stack. The validity of this pointer is complicated. The pointer can be used safely until the stack grows again and reaches the height at which the pointer was taken. However, it is not *guaranteed* that the pointer always refers to a live stack slot,   because a reallocation of the stack's backing array may occur.