@fjl 2022-05-11
I still think we should pursue this optimization idea, just need to find an abstraction that makes it less likely to mess things up.