@qianbin 2022-05-12
`popptr` is somewhat equivalent to `peek` & `pop`. Yes, it's dangerous especially for the stack with a slice backend. Agree that we need a better abstraction, I'll try.