@holiman 2022-05-12
It's actually _not_ terribly dangerous (see https://github.com/ethereum/go-ethereum/pull/24860#pullrequestreview-970350720). When something is pushed to the stack, the item is copied, so the stack owns it. 
The `peek` operation already returns the address. 

> returns a pointer to the 'inaccessible' area of the stack. 

As long as the caller only uses it in a very small scope, it is safe, just like the `peek` operation. However, for `peek` it's kind of obvious that this is a borrowed item which we cannot hold on to, which is not as apparent with `popptr`. So I agree that it needs to be made more obvious that this operation must not be used for anything longer than a single-op duration.