@holiman 2023-02-15
I don't think we want to merge this, since it doesn't markedly improve speed, but does make the stack interaction less simple